requirements.txt
requests==2.28.1
simplejson==3.16.0
jsonschema==3.2.0
pandas==1.3.5
datetime-truncate==1.1.1
parsedatetime==2.4
prometheus-client==0.15.0
rfc3339==6.2
python-dotenv==0.20.0
influxdb-client==1.32.0
